Fix wrong type signature

+++ clinica/pipelines/t1_volume/t1_volume_cli.py
@@ -38,7 +38,7 @@ def cli(
     dartel_tissues: List[int] = (1, 2, 3),
     tissues: List[int] = (1, 2, 3),
     modulate: bool = True,
-    voxel_size: Tuple[int, int, int] = (1.5, 1.5, 1.5),
+    voxel_size: Tuple[float, float, float] = (1.5, 1.5, 1.5),
     subjects_sessions_tsv: Optional[str] = None,
     working_directory: Optional[str] = None,
     n_procs: Optional[int] = None,
